+++ src/resolvers/index.js
@@ -5,6 +5,7 @@ const { mergeResolvers } = require('@graphql-tools/merge')
 module.exports = mergeResolvers([
 	require('./tokens'),
 	require('./permanentTokens'),
+	require('./contants'),
 	require('./records'),
 	require('./domains'),
 	require('./events'),