+++ src/stages/matchLimit.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { DURATIONS_LIMIT } = require('../constants/durations')
+const constants = require('../constants/durations')
 
 module.exports = () => {
 	// Some visitors keep sites open in the background. Their duration is often
@@ -8,7 +8,8 @@ module.exports = () => {
 	return {
 		$match: {
 			duration: {
-				$lt: DURATIONS_LIMIT,
+				// get the latest value using Proxy object
+				$lt: constants.DURATIONS_LIMIT,
 			},
 		},
 	}