+++ src/ui/scripts/api/fragments/constantFields.js
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client'
+
+export default gql`
+	fragment constantFields on Constants {
+		id
+        value
+        unit
+	}
+`