+++ src/ui/scripts/components/Label.js
@@ -8,7 +8,7 @@ const Label = (props) => {
 			className: classNames({
 				'label': true,
 				'label--no-spacing': props.spacing === false,
-			}),
+			}, props.className),
 			htmlFor: props.htmlFor,
 		}, props.children)
 	)
@@ -17,6 +17,7 @@ const Label = (props) => {
 Label.propTypes = {
 	spacing: PropTypes.bool,
 	htmlFor: PropTypes.string.isRequired,
+	className: PropTypes.string,
 	children: PropTypes.node.isRequired,
 }